+++ maturin/__init__.py
@@ -37,14 +37,17 @@ def get_maturin_pep517_args(
 ) -> list[str]:
     build_args = config_settings.get("build-args") if config_settings else None
     if build_args is None:
-        args = os.getenv("MATURIN_PEP517_ARGS", "")
-        if args:
+        env_args = os.getenv("MATURIN_PEP517_ARGS", "")
+        if env_args:
             print(
-                f"'MATURIN_PEP517_ARGS' is deprecated, use `--config-settings build-args='{args}'` instead."
+                f"'MATURIN_PEP517_ARGS' is deprecated, use `--config-settings build-args='{env_args}'` instead."
             )
+        args = shlex.split(env_args)
+    elif isinstance(build_args, str):
+        args = shlex.split(build_args)
     else:
         args = build_args
-    return shlex.split(args)
+    return args
 
 
 def _additional_pep517_args() -> list[str]: